Implement RDP text, keep within safe frame
dlists are now tracked independently of the RDP, I was having a lot of
trouble just waiting on RDP command start and end valid, or DMA busy, or
command ready. It seems like once you start reading this too often it
gets stuck? This might be an artifact of accessing it from the CPU.